split to separate files and improve test coverage
